Fix windows analysis

+++ projects/awinapi/last_error_message.cpp
@@ -13,17 +13,19 @@ export [[nodiscard]] auto last_error_message() -> std::string
 {
     DWORD error = ::GetLastError();
 
-    if (error == 0u)
+    if (error == 0U)
     {
         return {};
     }
 
     Buffer message;
 
-    message.tchar_size() = ::FormatMessage(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+    message.tchar_size() = ::FormatMessage(static_cast<DWORD>(FORMAT_MESSAGE_ALLOCATE_BUFFER)
+                                               | static_cast<DWORD>(FORMAT_MESSAGE_FROM_SYSTEM)
+                                               | static_cast<DWORD>(FORMAT_MESSAGE_IGNORE_INSERTS),
                                            nullptr,
                                            error,
-                                           MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
+                                           MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), // NOLINT(hicpp-signed-bitwise)
                                            static_cast<LPTSTR>(static_cast<void *>(&message.str())),
                                            0,
                                            nullptr);
